used replit to make portfolio. reorganized and deleted unnesesary files
I decided to attempt vibecoding. It is VERY powerful and made this website in less than an hour of promptiing. At first it made it with complicated lanusges and such (like react or express) yet all I wanted was a simple css, js and html project (i asked for that but it didn't hear me the first time I suppose.) I asked it to change that and it did but still hd some verbosity in files, structure and code. I had to delete most of it. now I am left with this and it is pretty good. I think I am going to use vide coding lightly in the future as the bug fixes and design choices are barely my own (although impressive.) I added the stars.js file from an old project as the vide coding couldn't give me the stars I wanted. wil fix bugs and content in future.